Moves behavioral interfaces to sub-namespace
