+++ src/Management/Behavior/Publishable.php
@@ -7,7 +7,7 @@
  * @license   MIT
  */
 
-namespace Contentful\Management;
+namespace Contentful\Management\Behavior;
 
 use Contentful\Management\Resource\SpaceScopedResourceInterface;
